Test cwl and cwl_cli separately
